+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/dd/DeleteDDCmd.java
@@ -15,10 +15,10 @@ import com.google.gson.Gson;
 
 import gov.nasa.pds.registry.common.es.client.EsClientFactory;
 import gov.nasa.pds.registry.common.es.client.EsUtils;
+import gov.nasa.pds.registry.common.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.Constants;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
 import gov.nasa.pds.registry.mgr.dao.RegistryRequestBuilder;
-import gov.nasa.pds.registry.mgr.util.CloseUtils;
 
 
 /**
