+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/reg/DeleteRegistryCmd.java
@@ -8,10 +8,10 @@ import org.elasticsearch.client.RestClient;
 
 import gov.nasa.pds.registry.common.es.client.EsClientFactory;
 import gov.nasa.pds.registry.common.es.client.EsUtils;
+import gov.nasa.pds.registry.common.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.Constants;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
 import gov.nasa.pds.registry.mgr.dao.IndexDao;
-import gov.nasa.pds.registry.mgr.util.CloseUtils;
 
 
 /**
