+++ src/main/java/gov/nasa/pds/registry/mgr/util/file/FileDownloader.java
@@ -25,7 +25,8 @@ import org.apache.http.ssl.TrustStrategy;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import gov.nasa.pds.registry.mgr.util.CloseUtils;
+import gov.nasa.pds.registry.common.util.CloseUtils;
+
 
 /**
  * File downloader with retry logic.
